Move state-related utils to salt.utils.state.py
Moves the following functions from `salt.utils` to `salt.utils.state`:
- gen_state_tag
- search_onfail_requisites
- check_onfail_requisites
- check_state_result

Original functions have a deprecation warning added and all references
in Salt code were updated to use the new function path, including moving
some unit tests.
