+++ salt/runners/state.py
@@ -10,6 +10,7 @@ import logging
 import salt.loader
 import salt.utils
 import salt.utils.event
+import salt.utils.state
 from salt.exceptions import SaltInvocationError
 
 LOGGER = logging.getLogger(__name__)
@@ -81,7 +82,7 @@ def orchestrate(mods,
             pillar_enc=pillar_enc,
             orchestration_jid=orchestration_jid)
     ret = {'data': {minion.opts['id']: running}, 'outputter': 'highstate'}
-    res = salt.utils.check_state_result(ret['data'])
+    res = salt.utils.state.check_state_result(ret['data'])
     if res:
         ret['retcode'] = 0
     else:
