+++ salt/states/pkg.py
@@ -81,9 +81,9 @@ import os
 import re
 
 # Import Salt libs
-import salt.utils  # Can be removed once gen_state_tag is moved
 import salt.utils.pkg
 import salt.utils.platform
+import salt.utils.state
 import salt.utils.versions
 from salt.output import nested
 from salt.utils import namespaced_function as _namespaced_function
@@ -3071,7 +3071,7 @@ def mod_aggregate(low, chunks, running):
     if low.get('fun') not in agg_enabled:
         return low
     for chunk in chunks:
-        tag = salt.utils.gen_state_tag(chunk)
+        tag = salt.utils.state.gen_state_tag(chunk)
         if tag in running:
             # Already ran the pkg state, skip aggregation
             continue
