+++ salt/utils/__init__.py
@@ -19,9 +19,7 @@ import fnmatch
 import hashlib
 import json
 import logging
-import numbers
 import os
-import posixpath
 import random
 import re
 import shlex
@@ -29,10 +27,8 @@ import shutil
 import socket
 import sys
 import pstats
-import tempfile
 import time
 import types
-import warnings
 import string
 import subprocess
 import getpass
@@ -40,7 +36,6 @@ import getpass
 # Import 3rd-party libs
 from salt.ext import six
 # pylint: disable=import-error
-from salt.ext.six.moves.urllib.parse import urlparse  # pylint: disable=no-name-in-module
 # pylint: disable=redefined-builtin
 from salt.ext.six.moves import range
 from salt.ext.six.moves import zip
@@ -134,7 +129,6 @@ from salt.exceptions import (
 
 
 log = logging.getLogger(__name__)
-_empty = object()
 
 
 def get_context(template, line, num_lines=5, marker=None):
@@ -1373,148 +1367,6 @@ def check_include_exclude(path_str, include_pat=None, exclude_pat=None):
     else:
         ret = True
 
-    return ret
-
-
-def gen_state_tag(low):
-    '''
-    Generate the running dict tag string from the low data structure
-    '''
-    return '{0[state]}_|-{0[__id__]}_|-{0[name]}_|-{0[fun]}'.format(low)
-
-
-def search_onfail_requisites(sid, highstate):
-    """
-    For a particular low chunk, search relevant onfail related
-    states
-    """
-    onfails = []
-    if '_|-' in sid:
-        st = salt.state.split_low_tag(sid)
-    else:
-        st = {'__id__': sid}
-    for fstate, fchunks in six.iteritems(highstate):
-        if fstate == st['__id__']:
-            continue
-        else:
-            for mod_, fchunk in six.iteritems(fchunks):
-                if (
-                    not isinstance(mod_, six.string_types) or
-                    mod_.startswith('__')
-                ):
-                    continue
-                else:
-                    if not isinstance(fchunk, list):
-                        continue
-                    else:
-                        # bydefault onfail will fail, but you can
-                        # set onfail_stop: False to prevent the highstate
-                        # to stop if you handle it
-                        onfail_handled = False
-                        for fdata in fchunk:
-                            if not isinstance(fdata, dict):
-                                continue
-                            onfail_handled = (fdata.get('onfail_stop', True)
-                                              is False)
-                            if onfail_handled:
-                                break
-                        if not onfail_handled:
-                            continue
-                        for fdata in fchunk:
-                            if not isinstance(fdata, dict):
-                                continue
-                            for knob, fvalue in six.iteritems(fdata):
-                                if knob != 'onfail':
-                                    continue
-                                for freqs in fvalue:
-                                    for fmod, fid in six.iteritems(freqs):
-                                        if not (
-                                            fid == st['__id__'] and
-                                            fmod == st.get('state', fmod)
-                                        ):
-                                            continue
-                                        onfails.append((fstate, mod_, fchunk))
-    return onfails
-
-
-def check_onfail_requisites(state_id, state_result, running, highstate):
-    '''
-    When a state fail and is part of a highstate, check
-    if there is onfail requisites.
-    When we find onfail requisites, we will consider the state failed
-    only if at least one of those onfail requisites also failed
-
-    Returns:
-
-        True: if onfail handlers suceeded
-        False: if one on those handler failed
-        None: if the state does not have onfail requisites
-
-    '''
-    nret = None
-    if (
-        state_id and state_result and
-        highstate and isinstance(highstate, dict)
-    ):
-        onfails = search_onfail_requisites(state_id, highstate)
-        if onfails:
-            for handler in onfails:
-                fstate, mod_, fchunk = handler
-                ofresult = True
-                for rstateid, rstate in six.iteritems(running):
-                    if '_|-' in rstateid:
-                        st = salt.state.split_low_tag(rstateid)
-                    # in case of simple state, try to guess
-                    else:
-                        id_ = rstate.get('__id__', rstateid)
-                        if not id_:
-                            raise ValueError('no state id')
-                        st = {'__id__': id_, 'state': mod_}
-                    if mod_ == st['state'] and fstate == st['__id__']:
-                        ofresult = rstate.get('result', _empty)
-                        if ofresult in [False, True]:
-                            nret = ofresult
-                        if ofresult is False:
-                            # as soon as we find an errored onfail, we stop
-                            break
-                # consider that if we parsed onfailes without changing
-                # the ret, that we have failed
-                if nret is None:
-                    nret = False
-    return nret
-
-
-def check_state_result(running, recurse=False, highstate=None):
-    '''
-    Check the total return value of the run and determine if the running
-    dict has any issues
-    '''
-    if not isinstance(running, dict):
-        return False
-
-    if not running:
-        return False
-
-    ret = True
-    for state_id, state_result in six.iteritems(running):
-        if not recurse and not isinstance(state_result, dict):
-            ret = False
-        if ret and isinstance(state_result, dict):
-            result = state_result.get('result', _empty)
-            if result is False:
-                ret = False
-            # only override return value if we are not already failed
-            elif result is _empty and isinstance(state_result, dict) and ret:
-                ret = check_state_result(
-                    state_result, recurse=True, highstate=highstate)
-        # if we detect a fail, check for onfail requisites
-        if not ret:
-            # ret can be None in case of no onfail reqs, recast it to bool
-            ret = bool(check_onfail_requisites(state_id, state_result,
-                                               running, highstate))
-        # return as soon as we got a failure
-        if not ret:
-            break
     return ret
 
 
@@ -3403,3 +3255,89 @@ def get_colors(use=True, theme=None):
         'Oxygen. This warning will be removed in Salt Neon.'
     )
     return salt.utils.color.get_colors(use=use, theme=theme)
+
+
+def gen_state_tag(low):
+    '''
+    Generate the running dict tag string from the low data structure
+
+    .. deprecated:: Oxygen
+    '''
+    # Late import to avoid circular import.
+    import salt.utils.versions
+    import salt.utils.state
+    salt.utils.versions.warn_until(
+        'Neon',
+        'Use of \'salt.utils.gen_state_tag\' detected. This function has been '
+        'moved to \'salt.utils.state.gen_state_tag\' as of Salt Oxygen. This '
+        'warning will be removed in Salt Neon.'
+    )
+    return salt.utils.state.gen_state_tag(low)
+
+
+def search_onfail_requisites(sid, highstate):
+    '''
+    For a particular low chunk, search relevant onfail related states
+
+    .. deprecated:: Oxygen
+    '''
+    # Late import to avoid circular import.
+    import salt.utils.versions
+    import salt.utils.state
+    salt.utils.versions.warn_until(
+        'Neon',
+        'Use of \'salt.utils.search_onfail_requisites\' detected. This function'
+        'has been moved to \'salt.utils.state.search_onfail_requisites\' as of '
+        'Salt Oxygen. This warning will be removed in Salt Neon.'
+    )
+    return salt.utils.state.search_onfail_requisites(sid, highstate)
+
+
+def check_onfail_requisites(state_id, state_result, running, highstate):
+    '''
+    When a state fail and is part of a highstate, check
+    if there is onfail requisites.
+    When we find onfail requisites, we will consider the state failed
+    only if at least one of those onfail requisites also failed
+
+    Returns:
+
+        True: if onfail handlers suceeded
+        False: if one on those handler failed
+        None: if the state does not have onfail requisites
+
+    .. deprecated:: Oxygen
+    '''
+    # Late import to avoid circular import.
+    import salt.utils.versions
+    import salt.utils.state
+    salt.utils.versions.warn_until(
+        'Neon',
+        'Use of \'salt.utils.check_onfail_requisites\' detected. This function'
+        'has been moved to \'salt.utils.state.check_onfail_requisites\' as of '
+        'Salt Oxygen. This warning will be removed in Salt Neon.'
+    )
+    return salt.utils.state.check_onfail_requisites(
+        state_id, state_result, running, highstate
+    )
+
+
+def check_state_result(running, recurse=False, highstate=None):
+    '''
+    Check the total return value of the run and determine if the running
+    dict has any issues
+
+    .. deprecated:: Oxygen
+    '''
+    # Late import to avoid circular import.
+    import salt.utils.versions
+    import salt.utils.state
+    salt.utils.versions.warn_until(
+        'Neon',
+        'Use of \'salt.utils.check_state_result\' detected. This function'
+        'has been moved to \'salt.utils.state.check_state_result\' as of '
+        'Salt Oxygen. This warning will be removed in Salt Neon.'
+    )
+    return salt.utils.state.check_state_result(
+        running, recurse=recurse, highstate=highstate
+    )
